Resolved changes
