+++ app/BlogSite/logic/apps.py
@@ -2,5 +2,4 @@ from django.apps import AppConfig
 
 
 class LogicConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'logic'
+    name = "Logic"
