+++ app/BlogSite/logic/routers/__init__.py
@@ -1,0 +1,4 @@
+__all__ = (
+    "get_posts",
+)
+from .posts import posts_fetcher as get_posts
