--- app/BlogSite/logic/routes/__init__.py
@@ -1,4 +1,0 @@
-from .posts import posts_fetcher
-__all__ = (
-    "posts_fetcher"
-)
